+++ src/class/config/class-get-options.php
@@ -116,6 +116,16 @@ class Get_Options extends Settings {
 	}
 
 	/**
+	 * Get the type methods available for persons
+	 *
+	 * @param int $number Optional: a number to turn into plural if needed
+	 * @return array<string, string>
+	 */
+	public static function get_list_person_methods( int $number = 1 ): array {
+		return parent::define_list_person_methods( $number );
+	}
+
+	/**
 	 * Get the type of people elements that are used for taxonomy
 	 *
 	 * @param int $number Optional: a number to turn into plural if needed
@@ -136,12 +146,25 @@ class Get_Options extends Settings {
 	}
 
 	/**
+	 * Get the type items elements that are used for taxonomy
+	 *
+	 * @param int $number Optional: a number to turn into plural if needed
+	 * @return array<string, string>
+	 */
+	public static function get_list_all_items( int $number = 1 ): array {
+		return [
+			...parent::define_list_non_taxo_items( $number ),
+			...parent::define_list_taxo_items( $number ),
+		];
+	}
+
+	/**
 	 * Get All taxonomy types: all people and items elements that are used for taxonomy
 	 *
 	 * @param int $number Optional: a number to turn into plural if needed
 	 * @return array<string, string>
 	 */
-	public static function get_list_all_elements_taxo( int $number = 1 ): array {
+	public static function get_list_fields_taxo( int $number = 1 ): array {
 		return [
 			...parent::define_list_taxo_people( $number ),
 			...parent::define_list_taxo_items( $number ),
@@ -176,7 +199,7 @@ class Get_Options extends Settings {
 	 * @param int $number Optional: a number to turn into plural if needed
 	 * @return array<string, string>
 	 */
-	public static function get_all_items( int $number = 1 ): array {
+	public static function get_all_fields( int $number = 1 ): array {
 		return [
 			...parent::define_list_non_taxo_items( $number ),
 			...parent::define_list_taxo_people( $number ), // Taxo_people is all people options, since there are no people options that are not taxonomy.
@@ -194,7 +217,7 @@ class Get_Options extends Settings {
 	 * @phpstan-return array{actor?: string, alsoknow?: string, connection?: string, goof?: string, plot?: string, producer?: string, quote?: string, soundtrack?: string, tagline?: string, trailer?: string, writer?: string}
 	 */
 	public static function get_items_with_numbers( int $number = 1 ): array {
-		$list_all = self::get_all_items( $number );
+		$list_all = self::get_all_fields( $number );
 		$list_elements_with_numbers = [];
 		$list_keys_with_numbers = array_keys( Settings::DATA_DEFAULT_WITHNUMBER );
 		foreach ( $list_all as $element => $translation ) {