+++ src/class/templates/admin/data/admin-data-taxonomy.php
@@ -126,7 +126,7 @@ if ( $lum_that->imdb_admin_values['imdbtaxonomy'] !== '1' ) {
 if ( $lum_that->imdb_data_values[ 'imdbtaxonomy' . $lum_key ] === '1' ) {
 	$lum_link_taxo_copy = add_query_arg( '_wpnonce_linkcopytaxo', wp_create_nonce( 'linkcopytaxo' ), $lum_current_admin_page . $lum_key );
 	$lum_file_in_stylesheet_path = get_stylesheet_directory() . '/' . Get_Options::LUM_THEME_TAXO_FILENAME_START . $lum_that->imdb_admin_values['imdburlstringtaxo'] . $lum_key . '.php';
-	$lum_translated_item = Get_options::get_all_items()[ $lum_key ];
+	$lum_translated_item = Get_options::get_all_fields()[ $lum_key ];
 
 	// No field to update found and no template to be updated found, offer to copy .
 	if ( count( $lum_fields_updated ) === 0 && is_file( $lum_file_in_stylesheet_path ) === false ) {